research/lab/pdf2text_debug/app/requirements.txt
flask
pdfminer.six